added processing for announcement controller

+++ src/main/java/com/apartmentseller/apartmentseller/services/impl/AnnouncementServiceImpl.java
@@ -2,6 +2,7 @@ package com.apartmentseller.apartmentseller.services.impl;
 
 import com.apartmentseller.apartmentseller.domain.Announcement;
 import com.apartmentseller.apartmentseller.dto.AnnouncementDto;
+import com.apartmentseller.apartmentseller.dto.UserDto;
 import com.apartmentseller.apartmentseller.repository.AnnouncementRepository;
 import com.apartmentseller.apartmentseller.services.AnnouncementService;
 import com.apartmentseller.apartmentseller.services.MapperService;
@@ -11,6 +12,7 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -36,15 +38,34 @@ public class AnnouncementServiceImpl implements AnnouncementService {
         return announcementDto;
     }
 
-    public AnnouncementDto updateAnnouncement(AnnouncementDto announcementFromDB, AnnouncementDto announcement) {
-        BeanUtils.copyProperties(announcement, announcementFromDB, "id");
-        Announcement announcementEntity = MapperService.INSTANCE.announcementDtoMapToAnnouncementEntity(announcementFromDB);
-        announcementRepository.save(announcementEntity);
-        return announcementFromDB;
+    public AnnouncementDto updateAnnouncement(long announcementId, AnnouncementDto announcement, UserDto currentUser) throws Exception {
+        return announcementRepository.findById(announcementId)
+                .map(announcementEntity -> {
+                    if(!ServiceUtils.hasUserPermissionToUpdate(announcementEntity.getAuthor().getId(), currentUser)){
+                        return null;
+                    }
+                    BeanUtils.copyProperties(announcement, announcementEntity, "id", "author", "creationTime");
+                    announcementRepository.save(announcementEntity);
+                    return announcement;
+                }).orElseThrow(Exception::new);
     }
 
-    public void deleteAnnouncement(AnnouncementDto announcementDto) {
-        Announcement announcementEntity = MapperService.INSTANCE.announcementDtoMapToAnnouncementEntity(announcementDto);
-        announcementRepository.delete(announcementEntity);
+    public void deleteAnnouncement(long announcementId, UserDto currentUser) {
+        announcementRepository.findById(announcementId)
+                .map(announcementEntity -> {
+                    if(!ServiceUtils.hasUserPermissionToUpdate(announcementEntity.getAuthor().getId(), currentUser)){
+                        return Optional.empty();
+                    }
+                    announcementRepository.delete(announcementEntity);
+                    return Optional.empty();
+                });
+
+//        Announcement announcementEntity = MapperService.INSTANCE.announcementDtoMapToAnnouncementEntity(announcementDto);
+//        announcementRepository.delete(announcementEntity);
+    }
+
+    @Override
+    public Optional<AnnouncementDto> getAnnouncement(long announcementId) {
+        return announcementRepository.findById(announcementId).map(MapperService.INSTANCE::announcementEntityMapToAnnouncementDto);
     }
 }
